refactor: deprecate old window classes
